Split frontend theme and Handlebars extensions into separate files
