+++ iris_core/modules/core/frontend/frontend.js
@@ -11,6 +11,7 @@ iris.registerModule("frontend");
 
 var fs = require('fs');
 var express = require('express');
+var path = require('path');
 
 /**
  *  Load theme
@@ -44,134 +45,9 @@ process.on("dbReady", function () {
 
 });
 
-var path = require("path");
+// Load in file for setting active theme
 
-/**
- * @function setActiveTheme
- * @memberof frontend
- *
- * @desc Sets the active name from the passed values.
- *
- * @param {string} themePath - path to the theme folder
- * @param {string} themeName - name of the active theme
- *
- * @returns error message if it fails.
- */
-iris.modules.frontend.globals.setActiveTheme = function (themeName) {
-
-  var glob = require("glob");
-
-  // Reset theme lookup registry
-
-  iris.modules.frontend.globals.templateRegistry.theme = [];
-
-  var result = {};
-
-  try {
-
-    var unmet = [];
-    var loadedDeps = [];
-
-    // Find theme
-
-    var found = glob.sync("{" + iris.rootPath + "/iris_core/themes/" + themeName + "/" + themeName + ".iris.theme" + "," + iris.sitePath + "/themes/" + themeName + "/" + themeName + ".iris.theme" + "," + iris.rootPath + "/home/themes/" + themeName + "/" + themeName + ".iris.theme" + "}");
-
-    var path = require("path");
-
-
-    if (found.length) {
-
-      var themeInfo = JSON.parse(fs.readFileSync(found[0]), "utf8");
-
-      var theme = {
-        name: themeName,
-        path: path.dirname(found[0]),
-        info: themeInfo
-      };
-
-    } else {
-
-      iris.log("error", "Could not find theme " + themeName);
-      return false;
-
-    }
-
-    // Read modules this theme is dependent on
-
-    if (theme.info.dependencies) {
-
-      Object.keys(theme.info.dependencies).forEach(function (dep) {
-
-        if (!iris.modules[dep]) {
-
-          unmet.push(dep);
-
-        }
-
-      })
-
-    }
-
-    // Push in theme templates to template lookup registry
-
-    if (!unmet.length) {
-
-      // Make config into a variable accessible by other modules
-
-      iris.modules.frontend.globals.activeTheme = theme;
-
-      iris.modules.frontend.globals.templateRegistry.theme.push(theme.path + "/templates");
-
-      // Push in theme's static folder
-
-      iris.app.use("/themes/" + themeName, express.static(theme.path + "/static"))
-
-    } else {
-
-      result.errors = "Active theme has unmet dependencies: " + unmet.join(",");
-
-    }
-
-  } catch (e) {
-
-    iris.log("error", e);
-
-    result.errors = "Something went wrong."
-
-  }
-
-  return result;
-
-}
-
-try {
-
-  var themeFile = fs.readFileSync(iris.sitePath + "/active_theme.json", "utf8");
-
-  try {
-
-    var activeTheme = JSON.parse(themeFile);
-
-    var setTheme = iris.modules.frontend.globals.setActiveTheme(activeTheme.name);
-
-    if (setTheme.errors) {
-
-      iris.log("error", "Could not enable " + activeTheme.name);
-      iris.log("error", setTheme.errors);
-
-    }
-
-  } catch (e) {
-
-    iris.log("error", e);
-
-  }
-
-} catch (e) {
-
-  iris.log("info", "No theme enabled");
-
-}
+require("./theme");
 
 /**
  * @function getTemplate
@@ -562,6 +438,8 @@ var findEmbeds = function (text, leaveCurlies) {
  * `context` array which is then present for rendering all templates. This means variables can be 
  * used across all templates when rendering.
  *
+ * TODO: Now embeds have been moved out into hook_frontend_embed this could be simplified
+ *
  * @param {string} html - HTML of template to process
  * @param {object} authPass - authPass of current user
  * @param {object} context - extra variables to pass to templating engine.
@@ -713,9 +591,12 @@ iris.modules.frontend.registerHook("hook_frontend_template_parse", 0, function (
         "src": "/modules/frontend/iris_core.js"
       },
       rank: 0
+
     }
 
-    data.variables["iris_theme"] = iris.modules.frontend.globals.activeTheme.name;
+    if (iris.modules.frontend.globals.activeTheme) {
+      data.variables["iris_theme"] = iris.modules.frontend.globals.activeTheme.name;
+    }
 
   }
 
@@ -916,110 +797,7 @@ iris.modules.frontend.registerHook("hook_display_error_page", 0, function (thisH
 
 });
 
-/**
- * @member hook_frontend_handlebars_extend
- * @memberof frontend
- *
- * @desc Template engine processing - Handlebars extending
- *
- * Attaches partials and extensions onto handlebars templating engine
- *
- * @returns Handlebars object
- */
-
-iris.modules.frontend.registerHook("hook_frontend_handlebars_extend", 0, function (thisHook, Handlebars) {
-
-  // Check route access
-
-  Handlebars.registerHelper("iris_menu", function (item) {
-
-    var route = iris.findRoute(item, "get");
-
-    if (route && route.options && route.options.permissions) {
-
-      if (iris.modules.auth.globals.checkPermissions(route.options.permissions, thisHook.authPass)) {
-
-        return item;
-
-      } else {
-
-        return null;
-
-      }
-
-    } else {
-
-      return item;
-
-    };
-
-  });
-
-  // Handle for a user's messages
-
-  Handlebars.registerHelper("iris_messages", function () {
-
-    var messages = iris.readMessages(thisHook.authPass.userid);
-
-    var output = "";
-
-    if (messages.length) {
-
-      output += "<ul class='iris-messages'>";
-
-      messages.forEach(function (message) {
-
-        output += "<li class='iris-message " + message.type + "'>" + message.message + "</li >";
-
-      });
-
-      output += "</ul>";
-
-      iris.clearMessages(thisHook.authPass.userid);
-
-    }
-
-    return output;
-
-  });
-
-  Handlebars.registerHelper('helperMissing', function (args) {
-
-    if (args.data.root.stripCurlies || args.data.root.finalParse) {
-
-      return "";
-
-    } else {
-
-      return "{{" + args.name + "}}";
-
-    }
-
-  });
-
-  Handlebars.registerHelper('iris_handlebars_delay', function (options) {
-
-    return options.fn();
-
-  });
-
-  Handlebars.registerHelper('iris_handlebars_ignore', function (options) {
-
-    if (options.data.root.finalParse) {
-
-      return options.fn();
-
-    } else {
-
-      return "{{{{iris_handlebars_ignore}}}}" + options.fn() + "{{{{/iris_handlebars_ignore}}}}"
-
-    }
-
-  });
-
-  thisHook.finish(true, Handlebars);
-
-});
+require("./handlebars_helpers")
 
 /**
  * @member hook_frontend_template
@@ -1152,11 +930,11 @@ var insertTags = function (html, vars) {
   }
 
   var embeds = findEmbeds(html);
-  
+
   var tags;
 
   if (embeds && embeds.tags) {
-    
+
     tags = embeds.tags;
 
   }
